% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalanche.R
\name{avalanche}
\alias{avalanche}
\title{Generate a class "avalanche" object}
\usage{
avalanche(series, year, rec_type, conf)
}
\arguments{
\item{series}{vector of series or tree names}

\item{year}{vector of years associated with each series and rec_type}

\item{rec_type}{vector of tree-ring record types. See \code{burnr} documentation}

\item{conf}{a category scaled 1-5 indicated the confidence assigned to a scar or wound as it relates to an avalanche event}
}
\value{
An \code{avalanche} and \code{fhx} object
}
\description{
Generate a class "avalanche" object
}
