% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.avalanche}
\alias{is.avalanche}
\title{Check if object is avalanche}
\usage{
is.avalanche(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean indicating whether `x` is an avalanche object.
}
\description{
Check if object is avalanche
}
